Bump to 4.0.0.0
PLEASE READ THE RELEASE NOTE FROM Andante CAREFULLY:
http://electronicobserver.blog.fc2.com/blog-entry-103.html

This release follows original EletronicObserver 4.0.0, which bundles a Browser implementation using CEF and will not work with .NET Framework < 4.5.2.

For EOE, this is a temporary release and we plan to separate the build / release of ElectronicObserver and EOBrowser (hopefully with 2 choices, CEF and Edge).
This is targeting to finish before original ElectronicObserver bumps to another version.

None of the known replacement Browser works in net40. We will support ElectronicObserver only :(